floating point calculations

+++ constants.h
@@ -1,7 +1,7 @@
 #pragma once
 
-constexpr double PI     =  3.1415926535;
+constexpr float PI     =  3.1415926535;
 
-constexpr double EPS    =  1e-4;
-constexpr double MIN_T  =  1e-4;
-constexpr double MAX_T  =  1e20;
+constexpr float EPS    =  1e-4;
+constexpr float MIN_T  =  1e-4;
+constexpr float MAX_T  =  1e20;
