+++ ray.cpp
@@ -1,14 +1,14 @@
 #include "ray.h"
 
 #include <cstdio>
-glm::mat3 orthonormalZ(const glm::dvec3& k){
-	using glm::dvec3;
+glm::mat3 orthonormalZ(const glm::vec3& k){
+	using glm::vec3;
 
-	/* dvec3 i = (std::abs(k.x)>EPS ? dvec3(0.0, 1.0, 0.0) : dvec3(1.0, 0.0, 0.0)); */
-	dvec3 i = dvec3(k.z-1.1, k.x, k.y);
+	/* vec3 i = (std::abs(k.x)>EPS ? vec3(0.0, 1.0, 0.0) : vec3(1.0, 0.0, 0.0)); */
+	vec3 i = vec3(k.z-1.1, k.x, k.y);
 	i = normalize(cross(i, k));
-	/* dvec3 i = rotz * rotx * k; */
-	const dvec3 j = cross(k, i);
+	/* vec3 i = rotz * rotx * k; */
+	const vec3 j = cross(k, i);
 	/* if(dot(k, i) > 0.1){ */
 	/*	   printf("k:%lf %lf %lf\n", k.x, k.y, k.z); */
 	/*	   printf("i:%lf %lf %lf\n", i.x, i.y, i.z); */
@@ -18,6 +18,6 @@ glm::mat3 orthonormalZ(const glm::dvec3& k){
 	return glm::mat3(i, j, k);
 }
 
-glm::dvec3 normalFromUV(const glm::dvec3& u, const glm::dvec3& v){
+glm::vec3 normalFromUV(const glm::vec3& u, const glm::vec3& v){
 	return normalize(cross(u, v));
 }
